requirements.txt
google-cloud-pubsub==2.18.0
apache-beam[gcp]==2.50.0
google-cloud-bigquery==3.13.0